creation of calculation rasters to own function
